==> scripts/01f-sunrise-sunset-times.ipynb <==
# Sunlight data

This script generates a graph of sunrise/sunset throughout the year,
for the 2nd presentation.

%pip install -r requirements.txt

import datetime as dt
import sys
import os

from astral.sun import sun
from astral.geocoder import database, lookup
from astral import LocationInfo
import pandas as pd

assert sys.version_info >= (3, 6), "Python version too low."

base_dir = "../" # parent of this directory
output_path = os.path.join(base_dir, "data/01f-sunrise.csv")

regions = {
    'QLD1': 'Brisbane',
    'NSW1': 'Sydney',
    'VIC1': 'Melbourne',
    'SA1': 'Adelaide',
    'TAS1': 'Hobart'
}

data = []

for (region, city_name) in regions.items():
    city_obj = lookup(city_name, database())
    
    d = dt.date(2009, 1, 1)
    
    while d.year < 2025:
        s1 = sun(city_obj.observer, date=d, tzinfo=f'Australia/{city_name}')
        s2 = sun(city_obj.observer, date=d, tzinfo='Australia/Brisbane')
        data.append({
            'city': city_name,
            'region': region,
            'd': d,
            'sunrise_local': s1['sunrise'],
            'sunset_local': s1['sunset'],
            'sunrise_fixed': s2['sunrise'],
            'sunset_fixed': s2['sunset'],
        })
        d = d + dt.timedelta(days=1)
df = pd.DataFrame(data)
df.head()

df.to_csv(output_path, index=False)
